ai-python/src/celery_worker_hub/import_worker/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
argcomplete==1.10.3
async-timeout==4.0.3
backports.zoneinfo;python_version<"3.9"
beautifulsoup4==4.8.2
billiard==4.2.0
boto3==1.34.60
botocore==1.34.60
celery==5.3.6
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
compressed-rtf==1.0.6
dnspython==2.6.1
docx2txt==0.8
ebcdic==1.1.1
email-validator==2.1.1
exceptiongroup==1.2.0
extract-msg==0.28.7
fastapi==0.110.0
flower==2.0.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
humanize==4.9.0
idna==3.6
IMAPClient==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
kombu==5.3.5
lxml==5.1.0
MarkupSafe==2.1.5
olefile==0.47
orjson==3.9.15
pdfminer.six==20191110
pillow==10.2.0
prometheus-client==0.20.0
prompt-toolkit==3.0.43
pycryptodome==3.20.0
pydantic==2.6.4
pydantic-core==2.16.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
s3transfer==0.10.0
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
starlette==0.36.3
textract==1.6.5
tornado==6.4
typing-extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
urllib3==1.26.18
uvicorn==0.28.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
xlrd==1.2.0
XlsxWriter==3.2.0
langchain==0.2.14
langchain-community==0.2.12
torch==2.2.2
transformers==4.39.2
sentence-transformers==2.6.1
httpx[http2]==0.27.0
pandas==2.2.1
tiktoken==0.6.0
pinecone-client[grpc]==3.2.2
rq==1.16.1
pymongo==4.6.3
PyMuPDF==1.24.9